requirements.txt
Django==4.2.9
inertia-django==0.5.3
django-vite==3.0.1
django-js-routes==0.2.0
python-decouple==3.8
django-import-export==3.3.5